requirements.txt
docker==6.0.1
GitPython==3.1.29
python-decouple==3.6
